+++ ort-custom-op/src/lib.rs
@@ -4,6 +4,7 @@ mod custom_op;
 mod error;
 mod inputs;
 mod outputs;
+mod value;
 
 pub mod prelude {
     pub use crate::api::{create_custom_op_domain, KernelInfo};
@@ -11,4 +12,5 @@ pub mod prelude {
     pub use crate::custom_op::{build, CustomOp};
     pub use crate::inputs::Inputs;
     pub use crate::outputs::Outputs;
+    pub use crate::value::Value;
 }